\name{cdid-package}
\alias{cdid-package}
\docType{package}
\title{Title of the cdid Package}
\description{
  Provide a detailed description of your package here.
}
\details{
  More detailed information on what the package does.
}
\author{
  List the authors of the package.
}
\seealso{
  Other related packages or functions.
}
\examples{
  # Example usage of your package
}
