% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{_PACKAGE}
\name{cdid}
\alias{cdid}
\title{cdid: An Overview of the cdid Package}
\description{
The cdid package contains functions for performing causal inference using Difference-in-Differences (DiD) methods with several advanced options for analysis.
}
