% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.aggte.R
\name{compute.aggte}
\alias{compute.aggte}
\title{Compute Aggregated Treatment Effect Parameters}
\usage{
\method{compute}{aggte}(
  MP,
  type = "group",
  balance_e = NULL,
  min_e = -Inf,
  max_e = Inf,
  na.rm = FALSE,
  bstrap = NULL,
  biters = NULL,
  cband = NULL,
  alp = NULL,
  clustervars = NULL,
  call = NULL
)
}
\arguments{
\item{call}{The function call to aggte}
}
\value{
\code{\link{AGGTEobj}} object
}
\description{
Does the heavy lifting on computing aggregated group-time
average treatment effects
}
\keyword{internal}
