% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_delta_to_att.R
\name{convert_delta_to_att}
\alias{convert_delta_to_att}
\title{convert_delta_to_att}
\usage{
convert_delta_to_att(dp)
}
\value{
a \code{\link{DIDparams}} object
}
\description{
Function to process arguments passed to the main methods in the
\code{cdid} package to compute ATT.
}
