% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_result.R
\name{convert_result}
\alias{convert_result}
\title{convert_result}
\usage{
convert_result(dp)
}
\value{
a \code{\link{DIDparams}} object
}
\description{
Result must be converted to be used in the aggte function.  .
}
