% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg}
\alias{gg}
\title{gg}
\usage{
gg(x, thet)
}
\arguments{
\item{x}{predictors}

\item{thet}{parameters}
}
\value{
A numeric vector representing the computed weights based on the predictors and parameters.
}
\description{
Function to simplify weight computations. For more details on the methodology, see:
Bellego, Benatia, and Dortet-Bernadet (2024), "The Chained Difference-in-Differences",
Journal of Econometrics, https://doi.org/10.1016/j.jeconom.2023.11.002.
}
\examples{
predictors <- matrix(c(1, 2, 3, 4), ncol = 2)
parameters <- matrix(c(0.5, -0.5), ncol = 1)
gg(predictors, parameters)

}
