% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{gmm}
\alias{gmm}
\title{chained GMM}
\arguments{
\item{tname}{The name of the column containing the time periods}

\item{yname}{The name of the outcome variable}

\item{idname}{The individual (cross-sectional unit) id name}

\item{gname}{The name of the variable in data that contains the first period when a particular observation is treated. This should be a positive number for all observations in treated groups. It defines which "group" a unit belongs to. It should be 0 for units in the untreated group.}

\item{xformla}{A formula for the covariates to include in the model. It should be of the form ~ X1 + X2. Default is NULL which is equivalent to xformla=~1. This is used to create a matrix of covariates which is then passed to the 2x2 DID estimator chosen in est_method.}

\item{propensityformla}{A formula for the covariates to include in the calculation of the propensity scores for the weights. This should be used only if weight_assumption is not NULL. Must be panel data.}

\item{data}{The name of the data.frame that contains the data}

\item{anticipation}{The number of time periods before participating in the treatment where units can anticipate participating in the treatment and therefore it can affect their untreated potential outcomes}

\item{weightsname}{The name of the column containing the sampling weights. If not set, all observations have same weight.}

\item{weight_assumption}{This is used to specify the assumption about the weights. The default is NULL which means that the weights are assumed to be constant over time. The other options are "missing_trends" and "sequential_missing". The first option assumes that the weights are constant over time, but that there are missing trends in the data. The second option assumes that the weights are constant over time, but that there are missing observations in the data. The last option is "varying" which assumes that the weights are varying over time. This is not yet implemented.}

\item{alp}{The significance level, default is 0.05}

\item{bstrap}{Boolean for whether or not to compute standard errors using the multiplier bootstrap. If standard errors are clustered, then one must set bstrap=TRUE. Default is TRUE (in addition, cband is also by default TRUE indicating that uniform confidence bands will be returned. If bstrap is FALSE, then analytical standard errors are reported.}

\item{biters}{The number of bootstrap iterations to use. The default is 1000, and this is only applicable if bstrap=TRUE.}

\item{base_period}{Whether to use a "varying" base period or a "universal" base period. Either choice results in the same post-treatment estimates of ATT(g,t)'s. In pre-treatment periods, using a varying base period amounts to computing a pseudo-ATT in each treatment period by comparing the change in outcomes for a particular group relative to its comparison group in the pre-treatment periods (i.e., in pre-treatment periods this setting computes changes from period t-1 to period t, but repeatedly changes the value of t). A universal base period fixes the base period to always be (g-anticipation-1). This does not compute pseudo-ATT(g,t)'s in pre-treatment periods, but rather reports average changes in outcomes from period t to (g-anticipation-1) for a particular group relative to its comparison group. This is analogous to what is often reported in event study regressions. Using a varying base period results in an estimate of ATT(g,t) being reported in the period immediately before treatment. Using a universal base period normalizes the estimate in the period right before treatment (or earlier when the user allows for anticipation) to be equal to 0, but one extra estimate in an earlier period.}

\item{print_details}{Whether or not to show details/progress of computations. Default is FALSE.}

\item{pl}{Whether or not to use parallel processing. Not yet functionning.}

\item{cores}{The number of cores to use for parallel processing. Not yet in function.}

\item{debT}{The first period of treatment.}

\item{finT}{The last period of treatment.}

\item{deb}{The first period of the data.}

\item{fin}{The last period of the data.}

\item{select}{A vector of observations to select.}

\item{treated}{The name of the column containing the treatment indicator.}

\item{clustervars}{The name of the column containing the clustering variable. Default is NULL.}

\item{cband}{Boolean for whether or not to compute a uniform confidence band that covers all of the group-time average treatment effects with fixed probability \code{1-alp}.  In order to compute uniform confidence bands, \code{bstrap} must also be set to \code{TRUE}.  The default is \code{TRUE}.}
}
\value{
\item{att}{The average treatment effect on the treated}
}
\description{
The chained difference-in-differences, leverages the overlapping structure of many unbalanced panel data set.
}
\examples{
data=data_sim = fonction_simu_attrition(nbsimu = 1, theta2_alpha_Gg=0.2, lambda1_alpha_St=0.2, sigma_alpha=2, sigma_epsilon=0.5)
results=gmm(
                yname="Y1_chaine",
                tname="annee",
                idname="id",
                gname="annee_G",
                xformla=~X, 
                propensityformla=c("X"), 
                data=data,      
                anticipation=0,      
                weightsname=c("P_Y1_chaine"), 
                # weight_assumption="missing_trends", #Not yet implemented.
                weight_assumption=NULL,
                bstrap=FALSE, 
                biters=1000,
                debT=3,
                finT=8,
                deb=1,
                fin=8,
                select='select',
                treated='traite_G',
                pl=FALSE,
                cores=1,
                cband=TRUE,
                clustervars=NULL)
results


}
