% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_compute_delta_att.R
\name{gmm_compute_delta_att}
\alias{gmm_compute_delta_att}
\title{GMM_compute_delta_att}
\usage{
gmm_compute_delta_att(dp)
}
\arguments{
\item{call}{Function call to gg}
}
\value{
a \code{\link{DIDparams}} object
}
\description{
Function to process arguments passed to the main methods in the
\code{cdid} package to compute Delta ATT as well as conducting some tests to make sure
data is in proper format.
}
