% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP.R
\name{print.MP}
\alias{print.MP}
\title{print.MP}
\usage{
\method{print}{MP}(x, ...)
}
\arguments{
\item{x}{An \code{MP} object, representing the results of multi-period analysis.}

\item{...}{Additional arguments passed to \code{summary.MP}.}
}
\value{
No return value. This function is called for its side effects of
printing the summary of the \code{MP} object to the console.
}
\description{
Prints a summary of the results contained in an \code{MP} object.
This function calls \code{summary.MP} to display the details of the multi-period
analysis results in a user-friendly format.
}
\seealso{
\code{\link{summary.MP}}
}
