% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_attgt.R
\name{process_attgt}
\alias{process_attgt}
\title{Process Results from \code{\link[=compute.att_gt]{compute.att_gt()}}}
\usage{
process_attgt(attgt.list)
}
\arguments{
\item{attgt.list}{list of results from \code{\link[=compute.att_gt]{compute.att_gt()}}}
}
\value{
list with elements:
\item{group}{which group a set of results belongs to}
\item{tt}{which time period a set of results belongs to}
\item{att}{the group time average treatment effect}
}
\description{
Process Results from \code{\link[=compute.att_gt]{compute.att_gt()}}
}
\references{
Callaway, Brantly and Pedro H.C. Sant'Anna.  \"Difference-in-Differences with Multiple Time Periods.\" Journal of Econometrics, Vol. 225, No. 2, pp. 200-230, 2021. \doi{10.1016/j.jeconom.2020.12.001}, \url{https://arxiv.org/abs/1803.09015}
}
