% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg}
\alias{gg}
\title{gg}
\usage{
gg(x, thet)
}
\description{
Function to simplify weight computations. For more details on the methodology, see:
Bellego, Benatia, and Dortet-Bernadet (2024), "The Chained Difference-in-Differences",
Journal of Econometrics, https://doi.org/10.1016/j.jeconom.2023.11.002.
}
