% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_compute_delta_att.R
\name{gmm_compute_delta_att}
\alias{gmm_compute_delta_att}
\title{GMM_compute_delta_att}
\usage{
gmm_compute_delta_att(dp)
}
\arguments{
\item{dp}{a dp object}
}
\value{
a \code{\link{DIDparams}} object
}
\description{
Function to compute the delta ATT. For more details on the methodology, see:
Bellego, Benatia, and Dortet-Bernadet (2024), "The Chained Difference-in-Differences",
Journal of Econometrics, https://doi.org/10.1016/j.jeconom.2023.11.002.
}
