% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_convert_result.R
\name{gmm_convert_result}
\alias{gmm_convert_result}
\title{convert_result}
\usage{
gmm_convert_result(dp)
}
\value{
a \code{\link{DIDparams}} object
}
\description{
Result must be converted to be used in the aggte function.  .
}
