% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_cdid.R
\name{pre_process_cdid}
\alias{pre_process_cdid}
\title{Process \code{did} Function Arguments}
\usage{
pre_process_cdid(
  yname,
  tname,
  idname,
  gname,
  xformla = NULL,
  data,
  panel = TRUE,
  allow_unbalanced_panel,
  control_group = c("nevertreated", "notyettreated"),
  anticipation = 0,
  weightsname = NULL,
  alp = 0.05,
  bstrap = FALSE,
  cband = FALSE,
  biters = 1000,
  clustervars = NULL,
  est_method = "dr",
  base_period = "varying",
  print_details = FALSE,
  pl = FALSE,
  cores = 1,
  call = NULL,
  treated
)
}
\value{
a \code{\link{DIDparams}} object
}
\description{
Function to process arguments passed to the main methods in the
\code{did} package as well as conducting some tests to make sure
data is in proper format / try to throw helpful error messages.
}
